JTs Edits
Style/look changes with the css and image folder.
Moved login up a row and added links in banner.xsl.
Anonymizer - hadn't realised there was one already here. I made my own,
with some specifics in it. As noted, we will need to check a couple of
things once tested and work out a way to remove the eamp namespace
declaration.